requirements.txt
cffi==1.13.2
cryptography==2.8
Django==2.2
django-tinymce==2.8.0
-e git+https://github.com/heroku/kafka-helper@70aa539540e7d6517f0c3c1900617767ab198b6c#egg=kafka_helper
mysqlclient==1.4.6
pycparser==2.19
pytz==2019.3
six==1.13.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
